test(scorecard): scorecard tests for recording management
